requirements.txt
flask
tensorflow
numpy
pandas
matplotlib
seaborn
pillow
tqdm
scikit-learn